+++ apps/realtime/src/features/socket.ts
@@ -1,4 +1,4 @@
-import { ISocketMessage } from "@bitrock-town/types";
+import { ISocketMessage } from "@bitrock/types";
 import { WebSocketServer } from "ws";
 import { verifyToken } from "../middleware/verifyToken";
 
@@ -12,27 +12,36 @@ socket.on("connection", async (ws, req) => {
 
   const token = req.url?.split("token=")[1];
 
-  if (!token) {
-    console.log("Client disconnected due to bad server configuration");
-    return ws.close();
+  const isBitstream =
+    req.headers.origin?.includes("localhost:3003") ||
+    req.headers.origin?.includes("stream.bitrock.town");
+
+  if (!isBitstream) {
+    if (!token) {
+      console.log("Client disconnected due to bad server configuration");
+      return ws.close();
+    }
+
+    const user = await verifyToken(token);
+    if (!user) {
+      console.log("Client disconnected due to invalid token");
+      return ws.close();
+    }
   }
-
-  const user = await verifyToken(token);
-
-  if (!user) {
-    console.log("Client disconnected due to invalid token");
-    return ws.close();
-  }
-
-  console.log("User connected:", user);
 
   ws.on("message", (message) => {
     const socketMessage = JSON.parse(message.toString()) as ISocketMessage;
     if (socketMessage.event === "position") {
       broadcast(socketMessage);
       return;
+    } else {
+      console.info("Received message:", socketMessage);
+      socket.clients?.forEach((client) => {
+        if (client !== ws && client.readyState === 1) {
+          client.send(JSON.stringify(message));
+        }
+      });
     }
-    ws.send(`Received message: ${message.toString()}`);
   });
 
   ws.on("close", () => {